requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
